daemon: move peer `jsonrpc_` methods to an individual module
